// Frontend/SalaryMark/src/app/shared/models/youtube/videos/item.ts
import { SafeResourceUrl } from '@angular/platform-browser';

export interface VideoItem {
  kind: string;
  etag: string;
  id: string;
  snippet: Snippet;
  url: SafeResourceUrl;
}

export interface Snippet {
  publishedAt: string;
  channelId: string;
  title: string;
  description: string;
  thumbnails: Thumbnails;
  channelTitle: string;
  playlistId: string;
  position: number;
  resourceId: ResourceID;
}

export interface ResourceID {
  kind: string;
  videoId: string;
}

export interface Thumbnails {
  default: Default;
  medium: Default;
  high: Default;
  standard: Default;
  maxres?: Default;
}

export interface Default {
  url: string;
  width: number;
  height: number;
}